+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-# v0.2.1
+# v0.2.4
 
 * Rename `.hao` into `.app`
 * Add executable comment to `hao.js`
 * Fix installation issue with missed links.
+* Rename `fs` provider into `local`.
+* Make local installer copy data to tmp folder
+* Add `beforeInstall`, `afterInstall`, `beforeUninstall` and `afterUninstall`
+  callbacks.